refactor: move files into module structure
