+++ src/app/modules/sns/sns.routes.ts
@@ -1,0 +1,18 @@
+import { Express } from 'express'
+
+import handleSns from './sns.controller'
+
+const mountSnsRoutes = (app: Express) => {
+  /**
+   * When email bounces, SNS calls this function to mark the
+   * submission as having bounced.
+   *
+   * Note that if anything errors in between, just return a 200
+   * to SNS, as the error code to them doesn't really matter.
+   *
+   * @route POST /emailnotifications
+   */
+  app.route('/emailnotifications').post(handleSns)
+}
+
+export default mountSnsRoutes
