+++ src/app/modules/sns/sns.types.ts
