+++ src/loaders/express/index.ts
@@ -7,6 +7,7 @@ import { Connection } from 'mongoose'
 import path from 'path'
 import url from 'url'
 
+import mountSnsRoutes from '../../app/modules/sns/sns.routes'
 import apiRoutes from '../../app/routes'
 import config from '../../config/config'
 
@@ -120,6 +121,7 @@ const loadExpressApp = async (connection: Connection) => {
   apiRoutes.forEach(function (routeFunction) {
     routeFunction(app)
   })
+  mountSnsRoutes(app)
 
   app.use(sentryMiddlewares())
